requirements.txt
asgiref==3.6.0
Bottleneck @ file:///opt/concourse/worker/volumes/live/f7139d11-2eee-4f6e-48cd-73168dc9e900/volume/bottleneck_1648028909796/work
brotlipy==0.7.0
cachetools==5.3.0
certifi==2020.6.20
cffi @ file:///Users/builder/ci_310/cffi_1643393868399/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cryptography @ file:///Users/builder/ci_310/cryptography_1642543758909/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Django==4.1.7
django-bootstrap5==22.2
fonttools==4.25.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
keras @ file:///home/ktietz/ci_310/keras_split_1643826818613/work/keras-2.6.0-py2.py3-none-any.whl
kiwisolver @ file:///Users/builder/ci_310/kiwisolver_1642533776499/work
matplotlib @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_croot-hkd3cnd2/matplotlib-suite_1647506470899/work
mkl-fft==1.3.1
mkl-random @ file:///opt/concourse/worker/volumes/live/a9eca864-069c-4240-73c8-4e51b036a10d/volume/mkl_random_1639994047876/work
mkl-service==2.4.0
munkres==1.1.4
numexpr @ file:///Users/builder/ci_310/numexpr_1642534542370/work
numpy @ file:///opt/concourse/worker/volumes/live/dd582496-15cc-4dc7-506b-bc24af13a378/volume/numpy_and_numpy_base_1649782811617/work
oauthlib==3.2.2
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.2
Pillow==9.0.1
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PySocks @ file:///Users/builder/ci_310/pysocks_1642536366386/work
python-crontab==2.7.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn @ file:///Users/builder/ci_310/scikit-learn_1642608976196/work
scipy @ file:///Users/builder/ci_310/scipy_1642604663510/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sqlparse==0.4.3
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tornado @ file:///Users/builder/ci_310/tornado_1642538510757/work
uritemplate==4.1.1
urllib3 @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_0b00bbfb-4d57-4b5b-af70-5677ec53f4ac495c5h1w/croots/recipe/urllib3_1650637219008/work
